+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Informative error when unable to find a post (in case it was deleted)
+
 ## [Version 1.1 (68)] - 2023-10-30T06:19:50Z
 
 ### Changes